+++ src/main/java/com/isocial/minisocialbe/repository/PostRepository.java
@@ -10,9 +10,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
     List<Post> findAllByContent(String content);
 
     @Query("SELECT DISTINCT p FROM Post p JOIN FETCH p.user u LEFT JOIN FETCH p.media m WHERE u.id = :userId")
-    List<Post> findByUserId(@Param("userId") Integer userId);
+    List<Post> findByUserId(@Param("userId") Long userId);
 }
